refactor(geo-tool): decompose extract.rs + realign bin into submodules (#1200)
TD-021 + TD-022. Behaviour-preserving file splits.

- extract.rs (954 LOC) -> extract/{mod,classify,dedup,crossroads}.rs.
  classify: OSM-tag classification + name generation; dedup: shared proximity
  de-dup; crossroads: junction extraction; mod.rs keeps the POI pass
  extract_features and re-exports classify_element/extract_crossroads so the
  super::extract::* paths in pipeline.rs are unchanged.
- realign bin (863 LOC) -> bin/realign_rundale_coords/{main,overrides,geocode,
  realign}.rs via the src/bin/<name>/main.rs layout (bin target name unchanged).
  overrides: --set-coord/--set-source/--baseline-world; geocode: Nominatim +
  suffix strip; realign: relative-position resolution + fictional realignment;
  main.rs keeps Cli, the shared WorldFile include!, and orchestration.

179 geo-tool tests pass (incl 21 realign + 22 extract); clippy -D warnings
clean; arch_fitness green. Update TODO.md + AGENTS.md module map.
