+++ parish/crates/parish-geo-tool/AGENTS.md
@@ -23,4 +23,4 @@ just realign-coords-run -- --baseline-world <path>            # realign from bas
 
 ## Module map
 
-`src/main.rs` CLI entry + `AdminLevel` enum, `src/bin/realign_rundale_coords.rs` coordinate realignment utility, `src/lib.rs` library surface (re-exports `osm_model`), `src/world_file_shared.inc` shared world-file serde types, `src/osm_model.rs` OSM data model types, `src/extract.rs` feature extraction from Overpass responses, `src/overpass.rs` Overpass API query builder + HTTP client, `src/merge.rs` merge OSM extracts with hand-authored data, `src/pipeline.rs` end-to-end download→extract→connect→describe→merge→output workflow, `src/lod.rs` level-of-detail density filtering, `src/output.rs` world.json output formatting, `src/descriptions.rs` location description template generation, `src/connections.rs` graph edge/connection building from road network, `src/cache.rs` HTTP response caching, `src/test_utils.rs` test helpers.
+`src/main.rs` CLI entry + `AdminLevel` enum, `src/bin/realign_rundale_coords/` coordinate realignment utility (`main.rs` CLI+orchestration, `overrides.rs`/`geocode.rs`/`realign.rs` submodules — #1200), `src/lib.rs` library surface (re-exports `osm_model`), `src/world_file_shared.inc` shared world-file serde types, `src/osm_model.rs` OSM data model types, `src/extract/` feature extraction from Overpass responses (`mod.rs` POI pass, `classify.rs`/`dedup.rs`/`crossroads.rs` submodules — #1200), `src/overpass.rs` Overpass API query builder + HTTP client, `src/merge.rs` merge OSM extracts with hand-authored data, `src/pipeline.rs` end-to-end download→extract→connect→describe→merge→output workflow, `src/lod.rs` level-of-detail density filtering, `src/output.rs` world.json output formatting, `src/descriptions.rs` location description template generation, `src/connections.rs` graph edge/connection building from road network, `src/cache.rs` HTTP response caching, `src/test_utils.rs` test helpers.
